feat: add moveImagesToPublicFolder function
This commit basically provides a better user experience when using
images in blog. Storing images in respective blog folder is a good
practice. But when user wants to use images in blog, he/she has to copy
the image to public folder.

This commit provides a function which will move the images to public
folder automatically.